Refactor employeeIds & customerIds by naming conventions
