+++ src/main/java/com/vinamaipo/hrm/controller/EmployeeController.java
@@ -1,7 +1,7 @@
 package com.vinamaipo.hrm.controller;
 
 import com.vinamaipo.hrm.domain.model.Employee;
-import com.vinamaipo.hrm.repos.EmployeeService;
+import com.vinamaipo.hrm.repository.EmployeeService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
